requirements.txt
discord.py
Lavalink.py
